+++ README.md
@@ -7,6 +7,8 @@ React SDK
 
 All the files within the template folder gets transpiled using [rollup](https://www.npmjs.com/package/rollup). Rollup helps bundling all dependencies and transpile them into CommonJS modules. This is required because this library will be used through NodeJS which does not understand these new modules natively and we do not want to limit the develop in which syntax they prefer.
 
+All transpiled files will be located in the folder `../__transpiled`.
+
 ## <a name="rendering_process"></a>The Rendering Process
