+++ src/index.ts
@@ -1,4 +1,5 @@
 export * from "./components";
 export * from "./renderer";
+export * from "./transpiler";
 export * from "./types";
 export { IndentationTypes } from "./utils";
